Moved BytesSize into types package
