+++ types/byte_size.go
@@ -1,4 +1,4 @@
-package config
+package types
 
 import (
 	"encoding/json"
